Fix return value type for cpio_scan
Last commit introduce a boolean return value
for cpio_scan. Previously, off_t was returned.
Fix type in both function and prototype.

+++ include/swupdate.h
@@ -110,6 +110,6 @@ struct swupdate_cfg {
 
 off_t extract_sw_description(int fd);
 off_t extract_next_file(int fd, int fdout, off_t offset, int compressed);
-off_t cpio_scan(int fd, struct swupdate_cfg *cfg, off_t start);
+int cpio_scan(int fd, struct swupdate_cfg *cfg, off_t start);
 
 #endif
